requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
future==0.18.2
h11==0.13.0
httptools==0.4.0
idna==3.3
oauthlib==3.2.0
python-dotenv==0.20.0
python-twitter==3.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sniffio==1.2.0
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3